split changes and reduce out-of-scope changes
